Split graph-storage lib.rs into smaller modules
