+++ document/graph-storage/src/ids.rs
@@ -1,0 +1,92 @@
+use crate::RegistryDelta;
+use serde::{Deserialize, Serialize};
+
+pub type NodeId = u64;
+pub type NetworkId = u64;
+/// Content-addressed identity for a `Delta`.
+/// 128-bit blake3 truncation: comfortable collision headroom for any plausible document lifetime
+/// without being adversarial-grade. Same delta content always produces the same `Rev`.
+pub type Rev = u128;
+
+/// Root network ID. The renderable graph lives in `networks[&ROOT_NETWORK]`.
+pub const ROOT_NETWORK: NetworkId = 0;
+
+/// Upper bound on a network's export slot count, guarding `SetExport` against a malicious or corrupted
+/// slot index forcing an unbounded `exports` allocation.
+pub(crate) const MAX_EXPORT_SLOTS: usize = 1 << 16;
+
+/// Unified storage-side position. The valid variants depend on `attr::UI_IS_LAYER`:
+/// layers use `Absolute` or `Stack`; non-layer nodes use `Absolute` or `Chain`.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub enum Position {
+	Absolute([i32; 2]),
+	Chain,
+	Stack(u32),
+}
+
+/// Per-device identity. Stable per `(device, document)`. Used for CRDT tiebreaking and `NodeId`
+/// scoping. Globally unique across all peers ever in a document.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Serialize, Deserialize)]
+pub struct PeerId(pub u64);
+
+/// Per-human identity. Stable across devices (one user, many devices). Used for identity display
+/// and undo-chain walking. Derived from `PeerId` via `Registry.peer_users`.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Serialize, Deserialize)]
+pub struct UserId(pub u64);
+
+/// Lamport timestamp with a peer-ID tiebreak. Higher counter wins; ties broken by peer.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Serialize, Deserialize)]
+pub struct TimeStamp {
+	pub counter: u64,
+	pub peer: PeerId,
+}
+
+impl TimeStamp {
+	/// Pre-edit origin. Used by initial `from_runtime` conversion before any edits have happened.
+	pub const ORIGIN: Self = TimeStamp { counter: 0, peer: PeerId(0) };
+}
+
+#[derive(Copy, Clone, Debug, Default, Serialize, Deserialize)]
+pub struct LamportClock {
+	pub(crate) counter: u64,
+	peer: PeerId,
+}
+
+impl LamportClock {
+	pub fn new(peer: PeerId) -> Self {
+		Self { counter: 0, peer }
+	}
+
+	/// Mints a fresh local timestamp.
+	pub fn tick(&mut self) -> TimeStamp {
+		self.counter += 1;
+		TimeStamp {
+			counter: self.counter,
+			peer: self.peer,
+		}
+	}
+
+	/// Advances past an incoming op so future local ticks are causally later.
+	pub fn observe(&mut self, incoming: TimeStamp) {
+		self.counter = self.counter.max(incoming.counter);
+	}
+}
+
+pub(crate) fn mint_node_id(peer: PeerId, counter: u64) -> NodeId {
+	let bytes = rmp_serde::to_vec(&(peer, counter)).expect("(PeerId, counter) must serialize");
+	let digest = blake3::hash(&bytes);
+	let mut truncated = [0u8; 8];
+	truncated.copy_from_slice(&digest.as_bytes()[..8]);
+	NodeId::from_le_bytes(truncated)
+}
+
+/// Hash the identity-bearing fields of a `Delta` with blake3 and truncate to 128 bits.
+pub(crate) fn compute_rev(parents: &[Rev], author: PeerId, timestamp: TimeStamp, delta_type: &RegistryDelta) -> Rev {
+	let mut hasher = blake3::Hasher::new();
+	let bytes = rmp_serde::to_vec(&(parents, author, timestamp, delta_type)).expect("Delta identity fields must serialize");
+	hasher.update(&bytes);
+	let digest = hasher.finalize();
+	let mut truncated = [0u8; 16];
+	truncated.copy_from_slice(&digest.as_bytes()[..16]);
+	Rev::from_le_bytes(truncated)
+}
